Feature/storypoints map details

+++ Maplify/MapClasses/MCMapDataSource.swift
@@ -17,19 +17,21 @@ protocol MCMapDataSourceDelegate {
 class MCMapDataSource {
     var mapView: MCMapView! = nil
     var mapService: MCMapService! = nil
-    var delegate: MCMapDataSourceDelegate! = nil
-
-    func reloadMapView() {
-        self.mapService.removeAllItems()
-        
-        let numberOfGroups: Int = (self.delegate?.numberOfGroups())!
-        
-        for i in 0...numberOfGroups {
-            let numberOfItems = (self.delegate?.numberOfMapItemsForGroup(i))!
-            for j in 0...numberOfItems {
-                let indexPath = NSIndexPath(forRow: i, inSection: j)
-                let mapItem = self.delegate?.mapItem(self.mapView, indexPath: indexPath)
-                self.mapService.placeItem(mapItem!)
+    var mapActiveModel: MCMapActiveModel! = nil
+    var delegate: AnyObject! = nil
+    
+    func reloadMapView<T: MCMapItem>(type: T.Type) {
+        if self.mapService != nil {
+            self.mapService.removeAllItems()
+            
+            for i in 0..<self.mapActiveModel.numberOfSections() {
+                for j in 0..<self.mapActiveModel.numberOfItems(i) {
+                    let indexPath = NSIndexPath(forRow: j, inSection: i)
+                    let data = self.mapActiveModel.cellData(indexPath)
+                    let mapItem = T()
+                    mapItem.configure(data)
+                    self.mapService.placeItem(mapItem)
+                }
             }
         }
     }
